Refactorings
- Simplify Checks to use Gauva Range and Preconditions
- Remove unnecessary toStrings()
- Inline some Map factory methods
- Refactor to use StopWatch instead of Clock for runtime timing
- Remove unused Loggables debug() and rename info() to log()
- Refactor BaseRange to inherit from AbstractDoubleCollection
- Simplify DescriptiveStats constructors and outliers()

+++ src/main/java/org/sharetrace/data/sampling/BaseRiskScoreSampler.java
@@ -12,7 +12,7 @@ abstract class BaseRiskScoreSampler extends BaseSampler<RiskScore> {
   @Override
   public RiskScore sample() {
     return RiskScore.builder()
-        .value((float) (RiskScore.VALUE_RANGE * normalizedSample(valueDistribution())))
+        .value(RiskScore.VALUE_RANGE * normalizedSample(valueDistribution()))
         .timestamp(timeSampler().sample())
         .build();
   }
